CI: Adjust required python version
Stupid restriction, but despite a project supporting e.g. python 3.3,
and only some secondary tools like sphinx needing more recent versions to run,
uv fails to install it due to specified `required-python` value.

Actually an argument to ditch uv, especially as all the CI is damn naive simple,
and uv seems over the top for it, but well. It's the current hype.